refactor(sdk): Support multiple contracts in `ChainEventPoller`  (#2928)
Refactored `ChainEventPoller` to support multiple contracts
simultaneously. Now the class is a singleton and we can define contract
methods as a listener parameter. This optimizes log fetching by reducing
required HTTP calls to the JSON RPC as there needs to be only one poller
running even if multiple contract events are listened.

Also we don't now need to create a `Contract` instance just for adding
event a listener. This allows us to simplify e.g. `Operator` class (in a
follow-up PR).