refactor(DHT): executeRecursiveOperation helper method in DhtNode (#3055)
## Summary

added new executeRecursiveOperation helper method to the network

## Future improvements

- The `deleteDataFromDht` and `findClosestNodesFromDht` should use the
function as well.
- Could make the creator parameter in storeDataToDht clearer. There
should be no need to provide it for the function. The External API could
simply be passed the storeManagers storeData method. Currently it is
used in the public API for testing but it should be reserved for
internal use only.