fix(node): Use `StreamrClient#getConfig()` to check WebRTC private address config (#3070)
## Background

The node checks a config value and prints "WebRTC private address
probing is disabled" if `webrtcAllowPrivateAddresses` is not true.

## Changes

Previously the client config was read directly from the config JSON. In
that approach the client defaults were not applied.
E.g. when using `"environment": "dev2"` node config option, it
incorrectly printed the warning (as `webrtcAllowPrivateAddresses` has a
default value of `true` in that environment).

Now it uses `StreamrClient#getConfig()`, and which contains the full
config with all defaults applied to it.