refactor(sdk): Operator contract utils without token ABI (#3081)
Refactored `delegate()` and `sponsor()` so that the token argument is
not needed. The token address is now queried via a
`Operator`/`Sponsorship` contract instance.

## Other changes

- removed obsolete `delegator` parameter from `transferAndCall()`
- simplified ABI definition in `getTestTokenContract()`
- added logging for `undelegate()`

## Future improvements

As we no longer need to pass the `token` argument for `delegate()` and
`sponsor()` functions it should be possible to promote the related
internal cli-tool command to non-internal commands.