feat(cli-tools): Internal `token-mint` command can mint for the authenticated user (#3080)
Improved internal `token-mint` command so that it easier to mint tokens
and gas for the authenticated user. Now it is possible to use keyword
`self` in place of `targetAddress`. When that is used, the target
address is calculated from the provided private key.

## Example

```
streamr internal token-mint self 1234 5678 --private-key ...
```

## Other changes

The permission commands also use special keyword similar like the
`self`-keyword introduced in this PR. Unified the cli-tools command help
to use the same style in those commands as we now use in the
`token-mint` command.