test: move `setupOperatorContract()` to `test-utils` (#3092)
Moved `setupOperatorContract()` utility from `sdk`'s
`operatorContractUtils` to `test-utils`. It is a test utility which
creates some test node wallets for an operator contract it deploys.

This move is a preparation for a PR in which we refactor
`operatorContractUtils` to be environment-independent.

## Other changes

Removed unnecessary `createTestWallet` option in
`checkOperatorValueBreach.test.ts`