feat(network): [NET-1451] Add option for Plum Tree like optimization of stream message propagation (#3083)
## Summary

Added support for Plum Tree like optimization of stream traffic. With
the mode enabled nodes propagate full messages only through the most
optimal path. The path is constructed based on duplicate messages
received by the node when it initially joins the stream. Nodes will
propagate the `MessageID`-field of the Stream Messages to any nodes that
are not on the optimal path. This allows reconstruction of the
propagation tree in the presence of churn.

## Changes

- Added `StreamPartDeliveryOptions` for streams that make it possible to
pass options of a per stream partition basis. This structure is used to
pass the option to use the plum tree optimization.

## TODO:

- More testing, unit tests

---------

Co-authored-by: Petri Savolainen <petri.savolainen@spaceify.net>