refactor(sdk): Move test utils from `operatorContractUtils` to `test-utils` (#3097)
There was some overlap between `sdk`'s `operatorContractUtils` and the
`test-utils` package. Now as the DATAv2 contract is available via
`@streamr/network-contracts` that overlapping can be avoided.

These functions should now be used from `test-utils` instead of
`operatorContractUtils`:
- `getTestTokenContract()`
- `getTestAdminWallet()`
- `getProvider()`

The `getProvider()` has been renamed to `getTestProvider()` and there is
stricter type for that.