refactor(network): Restructure trackerless-network directories (#3114)
## Summary

Restructured network directories. No longer use the outdated style from
the old network where most of the code is in a logic directory. Instead
code is split in directories based on the layer it is used in. If code
is multi-layer it is kept in the src directory.

## Future improvements

- Should code that is multilayer ie. NetworkStack,
ContentDeliveryManager, StreamPartReconnect have their own directory?

- Rename the `plum-tree` directory to something else that describes the
feature.