refactor: Use `waitForEvent()` instead of `waitForEvent3()` (#3122)
## Background

Before this PR there were separate functions of `waitForEvent()` and
`waitForEvent3()`, which has equal functionalities. The `waitForEvent()`
was for NodeJS's `EventEmitter` and `waitForEvent3()` was for the
`EventEmitter3` package.

## Changes

Refactored `waitForEvent()` to be a generic function so that it can be
used with both event emitters. Removed the redundant `waitForEvent3()`
function.

Refactored also `runAndWaitForEvents3()`, `raceEvents3()` and
`runAndRaceEvents3()` to be generic functions and renamed the suffixes
from the function names.

## Other changes

Added missing `IAutoCertifierClient#off()` function.

## Future improvements

Maybe we can simplify `runAndWaitForEvents3()`, `raceEvents3()` and
`runAndRaceEvents3()` as the functions now use `waitForEvent()`
internally, and therefore we could use e.g. `AbortController` there?