refactor(sdk): Required `earningsPerSecond` field in `deploySponsorshipContract()` (#3145)
Changed the `earningsPerSecond` to be a required field when calling
`deploySponsorshipContract()`. It is very important configuration field
for sponsorship deployment and therefore it makes sense that a value is
always provided.

Before this PR only tests relied on the default value (1 data token per
second). Maybe it is good to be explicit also in the tests, i.e. not
move the default value to from `deploySponsorshipContract()` to
`deployTestSponsorshipContract()` helper methods.