feat(network): [NET-1473] PlumTree can now be enabled from the sdk (#3147)
## Summary

It is now possible to enable the PlumTree optimization from the SDK.

## Changes

- renamed `plumTreeOptimisation` to `plumtreeOptimization`

## Future Improvements

- Perhaps `NetworkNode#broadcast` should not automatically join streams.
Instead when a message is published From the SDK for the first time
`joinStreamPart` should be called separately.
- Handle situation where user creates multiple subscriptions to the same
stream part with mismatching delivery options. An error could be thrown,
for example.
- Rename files in
`trackerless-network/src/content-delivery-layer/plum-tree` to be in line
with the `plumtreeOptimization` naming convention #3150