feat(sdk, cli-tools): [NET-1475] Unstake has `amount` parameter (#3154)
Added `amount` parameter to `operatorContractUtils#unstake(`). Also
added `dataTokenAmount` parameter to the related cli-tools command.

Before this PR unstaking removed the whole stake, i.e. it functioned
similarly how the `Operator` contract does internally. With this change
we have more usable interface and the interface in unified with the
`stake()` method.

## API change

Both the `operatorContractUtils` and the `operator-unstake` CLI tools
command are currently internal modules. Therefore this is not a breaking
change.