test(node): Update inspection smoke test (#3155)
Changes to the unstaking were needed because of these PRs:
- https://github.com/streamr-dev/network-contracts/pull/1014
- https://github.com/streamr-dev/network/pull/3154

In the first PR the addresses of pre-baked operators and sponsorships
changed. Therefore the operator and sponsorship contract addresses which
were listed in the test didn't exists anymore in `dev-chain-fast`. The
addresses of pre-baked operators/sponsorships are no longer hardcoded.
Instead the test queries the operator contract address from the
sponsorship factory, and fetches the list of sponsorships from The
Graph. The outcome is the same: all stakes done by the two test
operators are unstaked.

In the second PR the unstaking amount was added as a parameter. The
amount configured in the test didn't match the staked amount. Now the
staked amount is queried from the operator contract.