chore(sdk): [NET-1462] `StreamrClient#searchStreams()` queries using `Stream` entity (#3132)
Changed the `StreamrClient#searchStreams()` to use `Stream` entity
directly. The previous implementation used the `StreamPermission`
entity, which made the queries more complex and slower. See benchmarks
below.

This also fixes a potential bug as the previous implementation relied on
the contents of `StreamPermission#id` to be prefixed with stream id.
Most or all of the IDs in our subgraphs have used that kind of id
format, but it is not a documented feature of the subgraphs.

## Benchmarks

The benchmarks were run in `dev2` environment which had ~10000 streams

| Test | Before | After | Speed reduction
|---|---|---|---|
| query all streams | ~2700ms | ~480ms | ~80% |
| search one stream which matches a search term | ~44ms | ~34ms | ~25% |
| search permissions for user  | ~72ms | ~50ms | ~30% |
| search permissions for user, allowing public permission | ~200ms |
~100ms | ~50% |

## Related changes

Removed obsolete `createWhereClause()` method from `TheGraphClient`.