feat(node): Undelegation queue in Autostaker (#3163)
## Summary

Entries in the undelegation queue now affect the Autostaker
calculations.

## Background
 
When a delegator undelegates some amount from an operator, the tokens
aren't immediately transferred back to the delegator. Instead the amount
is added to the undelegation queue. The transfer happens as soon as
there are enough unstaked tokens in the operator contract.

## Changes

Modified the target stake calculation `payoutProportionalStrategy`. The
rounding logic and skipping of small transactions are also affected
slightly. The autostaker plugin reads the sum of undelegation queue
entries from The Graph and passes the amount to the strategy function.

Also added more assertions to the `autostaker.test.ts` smoke tests.

## Other changes

- The `myStakedAmount` in calculated from the stakes instead of querying
it from the contract
- Simplified logging in the plugin

## Chores

- Removed `sum.ts` from the `@streamr/utils` package
  - it was not exported
- the file was added during the development of
https://github.com/streamr-dev/network/pull/3086 (we should have removed
the file in that PR)
- Fixed field type in `StakeQueryResultItem`