feat(node) [NET-1485]: introduce more fine-grained excess token handling in autostaker algorithm (#3189)
## Summary

Introduce a more fine-grained excess token handling approach in the
stake/unstake planning algorithm. Compared to the previous approach,
this approach makes more of an effort to keep tokens staked at the cost
of added complexity.

Excess tokens appear during the algorithm when we filter out
stakings/unstakings from the plan that are deemed to small (their
adjustment is below `minTransactionAmount`). This can lead to there
being more tokens attempted to being staked than are available for
staking, a.k.a., excess tokens. Thus we need to balance this out by
adjusting the stakings of the plan.

### Previous approach in a nutshell

The previous approach looked for the staking with the smallest
adjustment. If the excess could be removed from this staking's
adjustment while still keeping the transaction valid* it would proceed
to do so. Otherwise it would remove the entire staking from the plan.
This process would repeat until there was no excess.

### New approach in a nutshell
The new approach will first try to remove the excess over all the
stakings. To do this, for each staking we calculate an _allowance_
(Finnish: vara, esim. turvavara), the amount we can reduce from the
adjustment while still keeping the transaction valid*. If the sum of
these allowances is more than or equal to the excess, we can safely
adjust the stakings and get rid of the excess without having to drop any
stakings. However, if the opposite is the case, we simply remove the
staking with the smallest adjustment from the plan (in similar vein to
the previous approach). This process is repeated until there is no
excess.

(*) a valid staking means the adjustment is above `minTransactionAmount`
and that the stake as a whole after the transaction is above
`minStakePerSponsorship`

## Limitations and future improvements

In the case when we do have to drop a staking from the plan (sum of
allowances isn't enough to cover excess), we often end up with unstaked
tokens at the end. We could add an extra step to the very end of the
algorithm, in which we attempt to divide any remaining tokens with the
stakings still present in the plan. However, this should be done in a
way as to minimize any potential oscillation.