feat: Support create-and-fund pattern in sponsorship deployment (#3230)
There is now an optional `sponsorAmount` option in
`deploySponsorshipContract()`. By using this option, deployment and
sponsoring can be performed in the same transaction.

## Changes

We now use `transferAndCall` instead of
`SponsorshipFactory#deploySponsorship()`. This allows us to pass both
pieces of information in a single transaction: the deployment options as
the `data` argument and the sponsor amount as the `value` argument.

If `sponsorAmount` is not defined, we simply transfer 0 wei. In that
case, a non-sponsored sponsorship is deployed, meaning the behavior is
identical to how it worked before this PR.