chore: Upgrade `eslint` project config for compatibility with `typescript-eslint` v8.46.4 (#3233)
Improved our config compatibility with `typescript-eslint` v8.46.4,
which introduced a parser error when both `parserOptions.project` and
`parserOptions.projectService` are set. See
https://github.com/typescript-eslint/typescript-eslint/releases/tag/v8.46.4

By removing the explicit project config and relying solely on
projectService, we avoid the error and remain fully compatible with this
release.

Also removed redundant parser definition (it is the default value in
`typescript-eslint`).

## Related PR

This is needed for https://github.com/streamr-dev/network/pull/3231