feat: `requiredBlockNumber` callback parameter in `TheGraphClient`  (#3238)
`TheGraphClient#queryEntities` callback now has extra parameter which
delivers the information about required block number.

We already had a poller in TheGraphClient, but it seems that that
approach doesn't always work (maybe the latest block number and the
actual query are be fetched from different indexers, maybe because there
is a load balancer in The Graph).

Also modified Autostaker to use that feature in the queries.

## Future improvements

- Implement the support also in `TheGraphClient#queryEntity` (which
doesn't currently use callbacks)