feat [NET-1639]: Bidirectional proxy connections (#3260)
## Summary

Added support for opening bidirectional proxy connections to the
network. These proxy connections can be used to publish and subscribe to
messages in the network. The bidirectional connections are now the
default behavior. This is an important feature for proxied nodes in the
StreamrTV. Previously it was possible to only set unidirectional publish
or subscribe proxy connections. Support for opening publish or subscribe
only proxy connections is kept in the network.