fix(cli-tools): Value required for `--partition` in `stream subscribe` (#3263)
Users must provide a partition value when using the `--partition` flag.
If the value is omitted, the command now shows a descriptive error
message.

There are no changes to the optionality of the flag: it is still ok to
leave it out, in which case the command defaults to partition `0`, just
like before this PR.

## Error

Before this PR we got this error like this when using the flag without a
value:

```
Error: invalid streamPartition value: true
    at ensureValidStreamPartitionIndex (.../network/packages/utils/dist/src/partition.js:10:15)
```

Now we show standard `commander` error:
```
error: option '-p, --partition <partition>' argument missing
```