refactor: Unbind `Logger` from NodeJS `module` object (by type) (#3283)
This pull request updates the `Logger` utility to make its scoping more
flexible and less dependent on NodeJS `module` objects.[^1] The changes
allow the logger to be initialized with either a string or an object
containing an `id`, making it easier to use in different contexts beyond
just files or modules.

It is backward compatible.

### Changes

**Logger scope improvements:**

* Changed the `Logger` constructor to accept a `scope` parameter, which
can be either a string or an object with an `id` property, instead of a
`NodeJS.Module`. (`packages/utils/src/Logger.ts`)
* Updated the `Logger.createName` static method to handle the new
`scope` type, extracting the identifier appropriately whether it's a
string or an object. (`packages/utils/src/Logger.ts`)
* Introduced the `Scope` type alias for clarity and type safety.
(`packages/utils/src/Logger.ts`)

### Future steps

* `module` is a Node-specific object and is not available in ES and
browser making the logger hard to use in other environments without a
polyfill. Next steps would be to stop initializing `Logger` with
`module`, esp. in files targetting multiple environments – subject for a
separate PR.

[^1]: Extracted (with changes) from #3279.