feat(sdk): Explicit encryption keys (#3284)
## Summary

Make it possible to provide explicit encryption keys to the
StreamrClient. These keys will be used in a given stream for all
publishers as the encryption key. Setting the explicit keys disables the
group key exchange. Therefore, all participants in the stream need to
agree on what key is being used for encryption and decryption to work
properly.
 
## Changes

- Added new configuration field `keys` to the `encryption` section
- The `GroupKeyManager#fetchKey` returns explicitly defined keys if they
are defined
- `Publisher` initializes the `GroupKeyQueue` with the explicit keys if
they are defined