feat(sdk): Add content delivery layer buffer configurations (#3305)
## Background

There are two different buffers related to this PR. The propagation
buffer and connection buffers. The propagation buffer is used when a
node is broadcasting data but does not have neighbors to send the data
to. The connection buffer is used when a new connection is formed in the
network.

## Summary

We can now configure these two options for the content delivery layer:
- `streamPartitionMaxPropagationBufferSize` can be used to set the size
of the propagation buffer.
- `contentDeliveryBufferWhileConnecting` can be used to stop buffering
while a connection is being formed.

## Related changes

- renamed `doNotBufferWhileConnecting` to `bufferWhileConnecting` in the
`trackerless-network` and `dht` packages.

## Example use case

In a case where a publisher joins the network before subscribers the
publisher stores the data to a buffer as it cannot send it to anyone.
When a subscriber joins the content of the buffer is sent to the
subscriber first and after that the real-time data is sent. If we don't
want that the buffer content is sent then we can use set
`streamPartitionMaxBufferSize` to 0 to avoid propagation buffering.

Additionally if the opening of connections takes long it is possible to
disable buffering on connecting connections to avoid sending of old
messages by setting the `contentDeliveryBufferWhileConnecting` as false.

---------

Co-authored-by: Teo Gebhard <teo@streamr.com>